requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
--extra-index-url https://www.piwheels.org/simple

ads1015==1.0.0
    # via enviroplus
astral==3.2
    # via enviroplus
cffi==1.17.1
    # via sounddevice
enviroplus==1.0.2
    # via enviroplus-exporter (pyproject.toml)
font-roboto==0.0.1
    # via enviroplus
fonts==0.0.3
    # via enviroplus
gpiod==2.2.1
    # via
    #   enviroplus
    #   gpiodevice
    #   pms5003
gpiodevice==0.0.5
    # via
    #   enviroplus
    #   pms5003
i2cdevice==1.0.0
    # via
    #   ads1015
    #   ltr559
    #   pimoroni-bme280
ltr559==1.0.0
    # via enviroplus
numpy==2.1.2
    # via st7735
paho-mqtt==2.1.0
    # via enviroplus
pimoroni-bme280==1.0.0
    # via enviroplus
pms5003==1.0.1
    # via enviroplus
prometheus-client==0.21.0
    # via enviroplus-exporter (pyproject.toml)
pycparser==2.22
    # via cffi
pyserial==3.5
    # via pms5003
pytz==2024.2
    # via enviroplus
smbus2==0.4.3
    # via i2cdevice
sounddevice==0.5.1
    # via enviroplus
spidev==3.6
    # via st7735
st7735==1.0.0
    # via enviroplus
